requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.0.1
faiss-cpu==1.12.0
sentence-transformers==5.1.0
transformers==4.55.2
torch==2.8.0
numpy==2.3.2
pandas==2.3.2
scikit-learn==1.7.1
PyPDF2==3.0.1
orjson==3.10.18
